Fix lots of compiler warnings due to format specifiers
They were either inconsistent with 32-bit or 64-bit compilation

+++ IMBFlickrSession.m
@@ -166,7 +166,7 @@
 	// Example of a photo that can't be downloaded: THE DECEIVING title.
 	
 	//	load the specified page...
-	NSString* page = [NSString stringWithFormat:@"%d", node.page + 1];
+	NSString* page = [NSString stringWithFormat:@"%ld", (long)node.page + 1];
 	[arguments setObject:page forKey:@"page"];
 	
 	return arguments;
